Moving delete buttons, making sure user can't delete themselves
Simply a move of the deletion buttons in the show.blade.php files for user & group to move it out of the section which is dedicated to different attributes of the user. Also, adding an if statement to the code which displays the delete button so that a user cannot delete themselves.

+++ umbrella-irrigation/resources/views/users/show.blade.php
@@ -17,10 +17,11 @@
             <label for="permission">Permission</label>
             <p data-editable data-type="number" data-name="permission">{{ $user->permission }}</p>
         </div>
-        <div class="container col-5">
-            <a href="/users/delete/{{$user->id}}" class="btn btn-primary btn-block btn-lg">Delete</a>
-        </div>
         <hr>
-
+        @if($user != Auth::user())
+            <div class="container col-5">
+                <a href="/users/delete/{{$user->id}}" class="btn btn-primary btn-block btn-lg">Delete</a>
+            </div>
+        @endif
     </form>
 @endsection
